+++ test/deprecated/sequence/CMakeLists.txt
@@ -7,5 +7,3 @@ string(REPLACE ".py" "" TEST_OPS "${TEST_OPS}")
 foreach(TEST_OP ${TEST_OPS})
   py_test_modules(${TEST_OP} MODULES ${TEST_OP})
 endforeach()
-set_tests_properties(test_sequence_conv PROPERTIES TIMEOUT 120)
-set_tests_properties(test_sequence_pool PROPERTIES TIMEOUT 120)
