+++ test/sequence/test_sequence_expand.py
